refactor: グラフ描画のロジックをhooksに切り出した
